+++ NEWS.md
@@ -1,6 +1,6 @@
 # version 1.0-24
 
-* `st_graticule()` creates graticule by default in the geographic coordinate reference system of arguments `x` or `crs`, and only if nothing is found or provided in `OGC:CRS84`
+* if `datum` is missing in a call to `st_graticule()`, a graticule by default will try to use the geographic coordinate reference system of arguments `x` or `crs`; when nothing is found there it falls back to `OGC:CRS84` (WGS84).
 
 * the figure margins parameter `mar` can be specified in a call to `plot.sf()`; #2558